+++ .fatherrc.ts
@@ -1,6 +1,5 @@
 import { defineConfig } from 'father';
 
 export default defineConfig({
-  esm: { output: 'es' },
-  cjs: { output: 'lib', platform: 'browser' },
+  esm: { output: 'dist' }
 });