+++ CHANGELOG.md
@@ -1,5 +1,159 @@
 # Changelog
 
+## [Version&nbsp;2.0.0-alpha.6](https://github.com/arvinxx/react-layout-kit/compare/v2.0.0-alpha.5...v2.0.0-alpha.6)
+
+<sup>Released on **2025-07-19**</sup>
+
+#### 🐛 修复
+
+- Try again memo.
+
+<br/>
+
+<details>
+<summary><kbd>Improvements and Fixes</kbd></summary>
+
+#### What's fixed
+
+- Try again memo ([3b4dd0f](https://github.com/arvinxx/react-layout-kit/commit/3b4dd0f))
+
+</details>
+
+<div align="right">
+
+[![](https://img.shields.io/badge/-BACK_TO_TOP-151515?style=flat-square)](#readme-top)
+
+</div>
+
+## [Version&nbsp;2.0.0-alpha.5](https://github.com/arvinxx/react-layout-kit/compare/v2.0.0-alpha.4...v2.0.0-alpha.5)
+
+<sup>Released on **2025-07-19**</sup>
+
+#### 🐛 修复
+
+- Try again pure css.
+
+<br/>
+
+<details>
+<summary><kbd>Improvements and Fixes</kbd></summary>
+
+#### What's fixed
+
+- Try again pure css ([ec86c19](https://github.com/arvinxx/react-layout-kit/commit/ec86c19))
+
+</details>
+
+<div align="right">
+
+[![](https://img.shields.io/badge/-BACK_TO_TOP-151515?style=flat-square)](#readme-top)
+
+</div>
+
+## [Version&nbsp;2.0.0-alpha.4](https://github.com/arvinxx/react-layout-kit/compare/v2.0.0-alpha.3...v2.0.0-alpha.4)
+
+<sup>Released on **2025-07-18**</sup>
+
+#### 🐛 修复
+
+- Fix layout issue again with fixed emotion css.
+
+<br/>
+
+<details>
+<summary><kbd>Improvements and Fixes</kbd></summary>
+
+#### What's fixed
+
+- Fix layout issue again with fixed emotion css ([8d2c05f](https://github.com/arvinxx/react-layout-kit/commit/8d2c05f))
+
+</details>
+
+<div align="right">
+
+[![](https://img.shields.io/badge/-BACK_TO_TOP-151515?style=flat-square)](#readme-top)
+
+</div>
+
+## [Version&nbsp;2.0.0-alpha.3](https://github.com/arvinxx/react-layout-kit/compare/v2.0.0-alpha.2...v2.0.0-alpha.3)
+
+<sup>Released on **2025-07-18**</sup>
+
+#### 🐛 修复
+
+- Fix layout issue.
+
+<br/>
+
+<details>
+<summary><kbd>Improvements and Fixes</kbd></summary>
+
+#### What's fixed
+
+- Fix layout issue ([63f70bf](https://github.com/arvinxx/react-layout-kit/commit/63f70bf))
+
+</details>
+
+<div align="right">
+
+[![](https://img.shields.io/badge/-BACK_TO_TOP-151515?style=flat-square)](#readme-top)
+
+</div>
+
+## [Version&nbsp;2.0.0-alpha.2](https://github.com/arvinxx/react-layout-kit/compare/v2.0.0-alpha.1...v2.0.0-alpha.2)
+
+<sup>Released on **2025-07-18**</sup>
+
+#### 🐛 修复
+
+- Fix build.
+
+<br/>
+
+<details>
+<summary><kbd>Improvements and Fixes</kbd></summary>
+
+#### What's fixed
+
+- Fix build ([2641c3c](https://github.com/arvinxx/react-layout-kit/commit/2641c3c))
+
+</details>
+
+<div align="right">
+
+[![](https://img.shields.io/badge/-BACK_TO_TOP-151515?style=flat-square)](#readme-top)
+
+</div>
+
+## [Version&nbsp;2.0.0-alpha.1](https://github.com/arvinxx/react-layout-kit/compare/v1.9.2...v2.0.0-alpha.1)
+
+<sup>Released on **2025-07-18**</sup>
+
+#### ⚡ 性能优化
+
+- Migration to pure css.
+
+<br/>
+
+<details>
+<summary><kbd>Improvements and Fixes</kbd></summary>
+
+#### Performance improvements
+
+- Migration to pure css ([797afab](https://github.com/arvinxx/react-layout-kit/commit/797afab))
+
+</details>
+
+#### 💥 BREAKING CHANGES
+
+- it will has css effects
+
+<div align="right">
+
+[![](https://img.shields.io/badge/-BACK_TO_TOP-151515?style=flat-square)](#readme-top)
+
+</div>
+
 ### [Version&nbsp;1.9.2](https://github.com/arvinxx/react-layout-kit/compare/v1.9.1...v1.9.2)
 
 <sup>Released on **2025-07-18**</sup>